+++ src/main/java/edu/ncsu/monopoly/RailRoadCell.java
@@ -3,22 +3,22 @@ package edu.ncsu.monopoly;
 public class RailRoadCell extends Cell {
 	static private int baseRent;
 	static public String COLOR_GROUP = "RAILROAD";
-	static private int price;
+	static private double price;
 
 	public static void setBaseRent(int baseRent) {
 		RailRoadCell.baseRent = baseRent;
 	}
 
-	public static void setPrice(int price) {
+	public static void setPrice(double price) {
 		RailRoadCell.price = price;
 	}
 	
-	public int getPrice() {
+	public double getPrice() {
 		return RailRoadCell.price;
 	}
 
-	public int getRent() {
-		return RailRoadCell.baseRent * (int)Math.pow(2, owner.numberOfRR() - 1);
+	public double getRent() {
+		return RailRoadCell.baseRent * Math.pow(2, owner.numberOfRR() - 1);
 	}
 	
 	public void playAction() {
